refactor(rules/prefer-derive-more-over-thiserror): split into a directory module
The rule had grown to 672 lines in a single flat file — the largest
in the repo and well past the one-screenful threshold at which
`CLAUDE.md` calls for a directory-module split. Break it apart along
its natural seams, following the `macro_trailing_comma` precedent and
keeping the flat `.rs` entry for the lint declaration, registration
functions, and the `EarlyLintPass` driver:

- `config` — `Config`, the `PreferDeriveMoreOverThiserror` pass state
  (fields `pub(super)` so the descendant submodules can read them),
  and the `path_matches_thiserror` matcher.
- `scan` — the crate-wide two-pass `use` / `extern crate` alias
  collection (`AliasCollector`, `CollectPhase`, `walk_use_tree_for_aliases`),
  exposed as `collect_aliases`.
- `detect` — per-item derive / attribute / import matching.
- `emit` — the diagnostic-emission and `#[error(...)]`-walking helpers.

Pure code movement: no behaviour change, the ui fixture and its
stderr are untouched, and `just all` (incl. the `--locked` self-lint
and the generated-catalogue check) passes.
